Move AlltoallMethodType
